requirements.txt
streamlit>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.22.0
numpy>=1.26.0
yfinance>=0.2.40
fastmcp>=0.4.0
huggingface_hub>=0.23.0
